[Fix/#92] 자동로그인 뷰 전환 로직 오류 해결
