+++ Staccato-iOS/Staccato-iOS/App/Staccato_iOSApp.swift
@@ -18,15 +18,29 @@ struct Staccato_iOSApp: App {
     private let alertManager = StaccatoAlertManager()
     @StateObject private var homeViewModel = HomeViewModel()
     @StateObject private var mypageViewModel = MyPageViewModel()
+    @StateObject private var signInViewModel = SignInViewModel()
 
+    init() {
+        signInViewModel.checkAutoLogin()
+    }
+    
     var body: some Scene {
         WindowGroup {
-            HomeView()
-                .environment(navigationState)
-                .environment(alertManager)
-                .environmentObject(homeViewModel)
-                .environmentObject(mypageViewModel)
+
+            Group {
+                if signInViewModel.isLoggedIn {
+                    HomeView()
+                        .environmentObject(homeViewModel)
+                        .environmentObject(mypageViewModel)
+                } else {
+                    SignInView()
+                }
+            }
+            .environment(navigationState)
+            .environment(alertManager)
+            .environmentObject(signInViewModel)
         }
+
     }
 
 }
